feature: Improve on the GraphQL performance on timeout
The data-fetchers do not abort when a timeout occurs. This commit
improves on the situation and in the end causes the GraphQL instance
to exit with a timeout exception.

The "notprivacysafe.graphql" logger should be turned off to prevent
the TimeOut exception to be logged. The risk is little, since any
exception logged here also propagate to the otp code calling
'GraphQL.execute()'.