fix: Improve error handling and prevent OTP from going down when connecting to external http services.
The VehicleRentalServiceDirectoryFetcher went down with a IllegalSateException when the
http endpoint failed. Instead of returning null in some error-cases and throwing IOExceptions
in others the HttpUtils is changed to throw an IOException in all cases. This make it more
robust, and the checked exception forces the client to handle it.