Fix: Regression error in passThroughPoints in the Transmodel API
The list of ids inside passThroughPoints is allowed to be empty or null. We cannot
change this - that would be a breaking change. So, when the via search enforced this,
the API was not backward compatible anymore. This commit reverts the behavior and
just ignores the passThroughPoints if the list of ids is null or empty. This bug was
 introduced in PR #6084.